% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_shap.R
\name{install_shap}
\alias{install_shap}
\title{Install shap Python library}
\usage{
install_shap(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto".
It is passed to the \code{\link[reticulate]{py_install}} function form package `reticulate`.}

\item{conda}{Path to conda executable.
It is passed to the \code{\link[reticulate]{py_install}} function form package `reticulate`.}
}
\description{
Install shap Python library
}
\examples{
\dontrun{
  install_shap((method = "auto", conda = "auto")
}

}
