% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/plot_individual_variable_importance.R
\name{plot.individual_variable_effect}
\alias{plot.individual_variable_effect}
\title{Plot Generic for Break Down Objects}
\usage{
\method{plot}{individual_variable_effect}(x, id = 1, ylevel = NULL,
  ...)

\method{plot}{individual_variable_effect}(x, id = 1, ylevel = NULL,
  ...)
}
\arguments{
\item{x}{the model model of 'individual_variable_effect' class}

\item{id}{of observation. By default first observation is taken.}

\item{ylevel}{level of y. By default first level is taken.}

\item{...}{other parameters}

\item{class}{number of class if model predicts multiple classes}

\item{observation}{number of observation if more than one was provided to 'individual_variable_effect'}

\item{x}{an individual variable effect explainer produced with function `individual_variable_effect()`}

\item{...}{other explainers that shall be plotted together}
}
\value{
a ggplot2 object
}
\description{
Function 'plot.individual_variable_effect' plots variables effects plots.
}
