% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_drwhy.R
\name{theme_drwhy_colors}
\alias{theme_drwhy_colors}
\title{DrWhy Theme for ggplot objects}
\usage{
theme_drwhy_colors(n = 2)
}
\arguments{
\item{n}{number of colors for color palette}
}
\value{
theme for ggplot2 objects
}
\description{
DrWhy Theme for ggplot objects
}
