% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.individual_variable_effect}
\alias{plot.individual_variable_effect}
\title{Plot Generic for Break Down Objects}
\usage{
\method{plot}{individual_variable_effect}(x, ..., class = 1,
  observation = 1)
}
\arguments{
\item{x}{the model model of 'individual_variable_effect' class}

\item{...}{other parameters}

\item{class}{number of class if model predicts multiple classes}

\item{observation}{number of observation if more than one was provided to 'individual_variable_effect'}
}
\description{
Plot Generic for Break Down Objects
}
