% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_individual_variable_importance.R
\name{print.individual_variable_effect}
\alias{print.individual_variable_effect}
\title{Print Individual Variable Importances}
\usage{
\method{print}{individual_variable_effect}(x, ...)
}
\arguments{
\item{x}{an individual variable importance explainer created with the \code{\link{`individual_variable_importance`}} function.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Individual Variable Importances
}
\examples{
\dontrun{
library("shapper")
library("DALEX")
library("randomForest")
Y_train <- HR$status
x_train <- HR[ , -6]
x_train$gender <- as.numeric(x_train$gender)
set.seed(123)
model_rf <- randomForest(x = x_train, y = Y_train)
p_fun <- function(x, data){
  predict(x, newdata = data, type = "prob")
  }
  ive <- individual_variable_effect(x = model_rf, data = x_train,
                                    predict_function = p_fun,
                                    new_observation = x_train[1,])
 ive
}

}
